Add a ConnectorUri class to make using Connector easier.
The current approach to using Connector means that we can't easily
implement approaches for handling 401 & 30X fallback (and maybe
others) directly inside the Connector class.

By adding a class for the URI we are connecting to, we will be more
easily able to implement redirect-following and auth-only-on-request
strategies when making requests.

This was evolved from the SearchUri subclass inside CheckServerDialog
where it proved very useful, so we'll start there with our refactoring
real soon.  This should also greatly assist with implementing support
for Digest authentication.

Signed-off-by: Andrew McMillan <andrew@morphoss.com>