Implement support for Digest Authentication.
Completely refactor the way we do requests at the same time,
so we handle redirects inside the requestor object in a way
that is appropriate for (Cal)DAV.  Authentication is handled
similarly to redirect in response to the first 401 received
for each transaction.

Service discovery is also refactored significantly to use HEAD
requests for port testing, followed by PROPFIND requests for
basic discovery and a final OPTIONS request to attempt to
retrieve CalDAV support status only when absolutely needed.

Signed-off-by: Andrew McMillan <andrew@morphoss.com>