Fix various issues with AcalRequestor class.
It was not retrying with Digest authentication in various
situations, so now it should do so.  There is also some very
weird stuff happening with requests when entity.getContentLength()
returns -1 (presumably the response headers do not contain a
content-length header) and we have to workaround this crappiness
by reading the response into a String and creating a new
InputStream from that.

Also, if our regrex couldn't parse the Location Uri, use the Uri
class. Probably we should just use it all the time :-)