This crude way of limiting infinite loops to ("only") 10000 is sad.
But it's kind of the only way we can do it.  Well, OK, maybe we could
make sure that the start position always increased in the outer loop.
but we're stuck with the inner loop not having that virtue, and this
is a much easier test that really will be fine except in the almost
obscure situation of someone having 9000 or so overrides.

Anyone like that deserves a day off anyway!