Rewrite & revalidate componentparts splitter.
There were some problems handling components with CRLF line endings vs.
components with LF line endings.  This fixes them so that aCal is now
completely line-ending agnostic. It should also be marginally be faster.

This is complex low-level stuff, so I'll leave the debugging in but
commented out.  There is a unit test for this which the code does now
pass.

Further efficiency could probably be gained by getting rid of the
regular expression matching.