Refactoring the way collectionId and resourceId are passed around.
They are now stored in the VCalendar and the AcalRepeatRule in order
that they can be applied in the creation of CalendarInstance.  I'm not
totally happy with this, but refactoring them out took me almost full
circle.  I think this is better than it was, but I wonder if we need
a CalendarResource to handle this information, which extends Resource
and holds a reference to the VCalendar.