Refactor the DatabaseTableManager to deal with transactions more sanely.
There was a serious breach of the principal of least surprise here
when beginTransaction would open the database, or openDB would begin
a transaction offering a confusing interlocking set of possibilities.

This now moves things back to openDB()/closeDB() deal with the
database as you'd expect and beginTx/endTx do transactions, again
as you would expect.