Save updates to text fields, from Matt Krai
"When I create a new task, edit its summary, and then set its due date,
aCal clears the task's summary.

"When the due date button is pressed, onCreateDialog creates a dialog
that, when the date is set, sets the todo's due date and calls
updateLayout.  updateLayout calls todo.getSummary to get the task's
summary and passes the result to todoName.setText.  Since the todo's
summary hasn't been set yet, this clears the summary.

"When editing an event, this is prevented by setting the event's
summary, location, and description before opening a dialog.  Do the
same thing when editing tasks.

Signed-off-by: Andrew McMillan <andrew@morphoss.com>